+++ ptyrad/__init__.py
@@ -1,5 +1,5 @@
 # __init__.py is not strictly needed since Python 3.3 for the intepreter to recognize it as a package 
 # However, this is still included for clarity and for initialization purpose
 
-__version__ = "v0.1.0-beta2.5" # 2024.09.19
+__version__ = "v0.1.0-beta2.6-acc" # 2024.09.30
 print(f"\nPtyRAD version: {__version__}")